CRUD example completion using jdbc

+++ src/main/java/com/achs/jdbc/example/maven/CrudOperations.java
@@ -27,7 +27,7 @@ public class CrudOperations {
 
         Connection conn = null;
         Statement statement = null;
-        String sql = "";
+        String sql;
 
         try {
             conn = DriverManager.getConnection(
@@ -72,9 +72,9 @@ public class CrudOperations {
                             System.out.println("Enter faculty:");
                             faculty = scanner.nextLine();
                             sql = "INSERT INTO users (full_name, address, mobile_number, college, faculty)"
-                                    + " VALUES ('"+fullName+"',"
-                                    + "'"+address+"', '"+mobileNumber+"',"
-                                    + "'"+college+"', '"+faculty+"' )";
+                                    + " VALUES ('" + fullName + "',"
+                                    + "'" + address + "', '" + mobileNumber + "',"
+                                    + "'" + college + "', '" + faculty + "' )";
                             
                             int value = statement.executeUpdate(sql);
                             if(value == 1) {
@@ -82,16 +82,90 @@ public class CrudOperations {
                             }
                             break;
                         case "update":
+                            System.out.println("Update request!!");
+                            System.out.println("Enter the row id whose information you want to update:");
+                            int id = Integer.parseInt(scanner.nextLine());
+                            
+                            if(id > 0) {
+                                sql = "Select * from users where id = " + id;
+                                result = statement.executeQuery(sql);
+                                if(result.next()) {
+                                    do {
+                                        fullName = result.getString("full_name");
+                                        address = result.getString("address");
+                                        mobileNumber = result.getString("mobile_number");
+                                        college = result.getString("college");
+                                        faculty = result.getString("faculty");
+                                    } while(result.next());
+                                    
+                                    String tempFullName, tempAddress, tempMobileNumber, tempCollege, tempFaculty;
+                                    System.out.println("Enter full name:");
+                                    tempFullName = scanner.nextLine();
+                                    System.out.println("Enter address:");
+                                    tempAddress = scanner.nextLine();
+                                    System.out.println("Enter mobile number:");
+                                    tempMobileNumber = scanner.nextLine();
+                                    System.out.println("Enter college:");
+                                    tempCollege = scanner.nextLine();
+                                    System.out.println("Enter faculty:");
+                                    tempFaculty = scanner.nextLine();
+                                    
+                                    if(tempFullName.equals("")) {
+                                        tempFullName = fullName;
+                                    }
+                                    
+                                    if(tempAddress.equals("")) {
+                                        tempAddress = address;
+                                    }
+                                    
+                                    if(tempMobileNumber.equals("")) {
+                                        tempMobileNumber = mobileNumber;
+                                    }
+                                    
+                                    if(tempCollege.equals("")) {
+                                        tempCollege = college;
+                                    }
+                                    
+                                    if(tempFaculty.equals("")) {
+                                        tempFaculty = faculty;
+                                    }
+                                    
+                                    sql = "UPDATE users SET full_name='"+tempFullName+"', address='"+tempAddress+"', "
+                                        + "mobile_number = '"+tempMobileNumber+"', college = '"+tempCollege+"', "
+                                        + "faculty = '"+tempFaculty+"' WHERE id = " + id;
+
+                                    value = statement.executeUpdate(sql);
+                                    if(value == 1){
+                                        System.out.println("Record updated successfully.");
+                                    } else {
+                                        System.out.println("Couldn't update the record.");
+                                    }
+                                } else {
+                                    System.out.println("The data in the row " + id + " does not exist.");
+                                }
+                            } else {
+                                System.out.println("Invalid ID.");
+                            }
                             break;
                         case "delete":
                             System.out.println("Delete request!!");
                             System.out.println("Please enter the row id to delete the information.");
-                            int id = scanner.nextInt();
-                            sql = "DELETE FROM users where id = " + id ;
-                            value = statement.executeUpdate(sql);
-                            System.out.println(value);
-                            if(value == 1) {
-                                System.out.println("Record deleted successfully.");
+                            id = Integer.parseInt(scanner.nextLine());
+                            
+                            if(id > 0) {
+                                sql = "Select * from users where id = " + id;
+                                result = statement.executeQuery(sql);
+                                if(result.next()) {
+                                    sql = "DELETE FROM users where id = " + id ;
+                                    value = statement.executeUpdate(sql);
+                                    if(value == 1) {
+                                        System.out.println("Record deleted successfully.");
+                                    } else {
+                                        System.out.println("Unable to delete the data.");
+                                    }
+                                } else {
+                                    System.out.println("Invalid row id.");
+                                }
                             }
                             break;
                         default:
